open-router/pyproject.toml
[project]
name = "open-router"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agents>=1.4.0",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]
